fix: fixed bug creating table when model has no columns
If you wanted to create a table using only the ColumnExpr function and
not provide columns from the struct model. Then you would receive syntax
errors as the ColumnExpr columns always have a `,` prepended before them
even if no columns were appended before them. This makes sure that the
`,` is properly prepended before adding columns from the ColumnExpr. It
also evaluates the AUTO_INCREMENT feature from the provided formatter
dialect rather than the DB. This will make testing the AppendQuery
method easier in the future as the library evolves as it this method is
now _slightly_ less dependent on a DB connection to be called.