feat: manage database schema with bundb CLI
Here we introduce bundb auto migrate command to create and run migrations from the CLI.
It works by importing the Models from a pre-build user plugin, similarly to how Migrator
gets the list of migrations by being compiled together with the user's package.
The difference here is that bundb can be distributed as a standalone binary and only requires
users to set up the migrations/main.go file.