feat(bundb): provide experimental commands under bundb
Experimental commans are the good old [auto, migrate, rollback, etc.]
but adapted to import configuration from a pre-built plugin.

Plugins are known to be unstable and only sometimes work correctly,
so this feature is more of a proof of concept and we should probably
add the option to disable it entirely with build tags (for standalone bundb).

TODO:: bun init --plugin does not initialize a Go module in the
migrations/ directory. Currently one would need to manually run go mod init
and go mod tidy after bootstraping the project.