fix: handle driver.Valuer in getRealValue
This commit addresses an issue where the `getRealValue` function was
not correctly handling fields that implement the `driver.Valuer`
interface. The code now checks if the field implements
`driver.Valuer` and, if so, retrieves the underlying value. It then
checks the Kind of type to determine whether it can be
returned directly.

fix https://github.com/uptrace/bun/issues/1138